@@ -34,25 +34,37 @@
   "dependencies": {
     "babel-cli": "^6.11.4",
     "babel-core": "^6.13.2",
+    "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.11.0",
-    "babel-register": "^6.11.6",
+    "babel-register": "^6.18.0",
     "diff-match-patch": "^1.0.0",
     "electron-squirrel-startup": "^1.0.0",
     "jquery": "2.2.4",
+    "material-ui": "^0.17.4",
     "pouchdb-adapter-leveldb": "^6.1.0",
     "pouchdb-core": "^6.1.0",
     "pouchdb-quick-search": "^1.2.0",
-    "react": "15.1.0",
+    "react": "^15.4.2",
     "react-addons-css-transition-group": "^15.5.2",
     "react-addons-shallow-compare": "^15.5.2",
     "react-bootstrap": "^0.30.8",
     "react-bootstrap-grid": "^1.0.0",
-    "react-dom": "15.1.0",
-    "react-selectize": "^2.1.0"
+    "react-dom": "^15.4.2",
+    "react-scripts": "0.8.5",
+    "react-selectize": "^2.1.0",
+    "react-tap-event-plugin": "^2.0.1",
+    "webpack": "^1.14.0"
   },
   "devDependencies": {
+    "babel-core": "^6.24.1",
+    "babel-loader": "^7.0.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.16.0",
+    "babel-preset-react-hmre": "^1.1.1",
+    "babel-preset-react-optimize": "^1.0.1",
+    "babel-preset-stage-0": "^6.16.0",
+    "babel-register": "^6.18.0",
     "electron": "1.4.12",
-    "rimraf": "^2.5.4",
     "electron-winstaller": "^2.5.1",
     "gulp": "^3.9.1",
     "gulp-bower": "0.0.13",
@@ -66,7 +78,8 @@
     "gulp-rename": "^1.2.2",
     "gulp-sass": "^3.0.0",
     "gulp-util": "^3.0.6",
-    "main-bower-files": "^2.9.0"
+    "main-bower-files": "^2.9.0",
+    "rimraf": "^2.5.4"
   },
   "build": {
     "appId": "oa.autographa.lite",
