@@ -1,17 +1,27 @@
-const React = require('react')
-const ReactDOM = require('react-dom')
-const Constant = require("../util/constants");
-const Tabs = require('react-bootstrap/lib/Tabs');
-const Tab = require('react-bootstrap/lib/Tab');
-const session = require('electron').remote.session;
-const { dialog } = require('electron').remote;
-
+// <<<<<<< HEAD
+// const React = require('react')
+// const ReactDOM = require('react-dom')
+// const Constant = require("../util/constants");
+// const Tabs = require('react-bootstrap/lib/Tabs');
+// const Tab = require('react-bootstrap/lib/Tab');
+// const session = require('electron').remote.session;
+// const { dialog } = require('electron').remote;
+// =======
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Tabs from 'react-bootstrap/lib/Tabs';
+import Tab from 'react-bootstrap/lib/Tab';
+const session =  require('electron').remote.session;
+const Constant = require("../util/constants")
+import { dialog } from 'electron';
+import { remote } from 'electron';
 
 class BookList extends React.Component {
 	constructor(props) {
         super(props);
+        this.onItemClick = this.onItemClick.bind(this);
         this.state = { 
-            data: booksList,
+            data: Constant.booksList,
             chapterData:[],
             book: 1,
             currentBook: 1,
@@ -24,13 +34,16 @@ class BookList extends React.Component {
 		session.defaultSession.cookies.get({ url: 'http://book.autographa.com' }, (error, cookie) => {
 	    	if (cookie.length > 0) {
 	            var bookNo = cookie[0].value;
-	            console.log(cookie);
 	            this.setState({bookNo: bookNo});
             } else {
 				this.setState({book: '1'})
         	} 	
 	    });
 	    console.log(this.state.showModalBooks);
+	}
+
+	onItemClick(bookNo) {
+  		global.book = bookNo;
 	}
      
 	handleSelect(key) {
@@ -41,38 +54,58 @@ class BookList extends React.Component {
 		this.setState({activeTab:key});
 		var chap = [];
 		var bookChapter = bookCodeList[parseInt(global.book, 10) - 1]
-		var bookno = global.book
-		this.setState({currentBook: global.book})
-		global.bookName = bookCodeList[parseInt(bookno, 10) - 1]
-		var id = 'eng_udb' + '_' + bookCodeList[parseInt(bookno, 10) - 1]
+
+		let bookNo = global.book
+		if(!bookNo){
+			bookNo = 1;
+		}
+		this.setState({currentBook: bookNo})
+		var id = 'eng_udb' + '_' + bookCodeList[parseInt(bookNo, 10) - 1]
 		var getData = refDb.get(id).then(function(doc) {
 			 doc.chapters.forEach(function(ref_doc) {
-		    chap.push({ number: chapter });
-			})  	
+		    	chap.push({ number: chapter });
+			})
 			return chap
-			console.log(chap)
+		}).catch(function(err){
+			
 		})
-			getData.then((item) =>{
-			this.setState({chapterData:item})
+
+		getData.then((item) =>{
+			if(item  && item.length)
+				this.setState({chapterData:item})
 		})
 	}
 
 	render() {
      let close = () => this.setState({showModalBooks:false });
 	    return ( 
-	    <Tabs animation={false} activeKey={this.state.activeTab} onSelect={() =>this.goToTab((this.state.key == 1) ? 2 : 1)} id="noanim-tab-example">
+	    <Tabs animation={false} activeKey={this.state.activeTab} onSelect={() =>this.goToTab((this.state.activeTab == 1) ? 2 : 1)} id="noanim-tab-example">
 		    <Tab eventKey={1} title="Book" onClick={() => this.goToTab(2)}>
 			    <div className="wrap-center"></div>
 	            <div className="row books-li" id="bookdata">
 	                <ul id="books-pane">
-	                    <BookGroup result={this.state.data} currentBook = {this.state.currentBook} />
+	                    {
+	                    	this.state.data.map((item,index) =>{
+								return <li key={index}><a href="#" key={index}  onClick = { this.onItemClick.bind(this, index+1) } value={item} className={(index+1 == this.state.currentBook) ? 'link-active': ""}>{item}
+								</a></li>
+							})
+
+	                    }
+	                	
 	                </ul>
 	            </div>
 				<div className= "clearfix"></div>
             </Tab>
 		    <Tab eventKey={2} title="Chapters" > 
 		    	<div className="chapter-no">
- 					<ChapterList chapterData = { this.state.chapterData } currentchapter = {this.state.currentchapter} onModalClose = {this.state.onModalClose}/> 
+
+                		<ul id="chaptersList">
+                			{
+                				this.state.chapterData.map(function(row, i) {
+									return ( <li key={i}>{i+1}</li> );
+								})
+                			}
+                		</ul>
             	</div>
             </Tab>
   		</Tabs>
@@ -83,7 +116,8 @@ class BookList extends React.Component {
 var BookGroup = function(props) {
 	const BooksGroup = props.result.map((item,index) =>{
 		let _handleClick = this.onItemClick.bind(this, index+1);
-		return <li key={index}><a href="#" onClick={_handleClick } value={item} className={(index+1 == props.currentBook) ? 'link-active': ""}>{item}
+
+		return <li key={index}><a href="#" key={index} onClick={_handleClick } value={item} className={(index+1 == props.currentBook) ? 'link-active': ""}>{item}
 		</a></li>
 	})
 	return (
