@@ -1,11 +1,9 @@
 (function() {
   require("babel-register")({
     extensions: [".js", ".jsx"],
-    presets: ["react"]
+    presets: ['es2015', 'react']
   });
-  const ReactDOM = require('react-dom');
-  const React = require('react');
-  const remote = require('electron').remote;
+
   
   var App = {
     init: function() {
