@@ -1,5 +1,5 @@
-const React = require('react');
-const ReactDOM = require('react-dom');
+const React = require('react')
+const ReactDOM = require('react-dom')
 
 class Footer extends React.Component 
 {
